+++ _src/objects.ts
@@ -11,22 +11,31 @@ export const skipInterceptorsProperty: string = 'skipInterceptors';
 export class NaniumObject<T> {
 	static strictDefault: boolean = false;
 
-	constructor(data?: Partial<T>, genericTypes?: NaniumGenericTypeInfo, strict?: boolean) {
-		NaniumObject.init(this, data, genericTypes, strict);
+	constructor(data?: Partial<T>, genericTypes?: NaniumGenericTypeInfo, strict?: boolean, cloneDeep?: boolean) {
+		NaniumObject.init(this, data, genericTypes, strict, cloneDeep ?? true);
 	}
 
 	private static initObjectCore<T = any>(
 		plain: any,
 		constructorOrObject: (new (data?: Partial<T>) => any) | T,
 		globalGenericTypes?: NaniumGenericTypeInfo,
-		localGenericTypes?: LocalGenerics | ConstructorType,
-		strict?: boolean
+		localGenericTypes?: LocalGenerics | ConstructorType | ConstructorGetter,
+		strict?: boolean,
+		deepClone?: boolean,
+		plainParents: any[] = [],
 	): any {
 		let constructor: ConstructorType;
 		let result: T;
 		if (typeof constructorOrObject === 'function') {
-			constructor = constructorOrObject as ConstructorType;
-			result = new constructor();
+			if (!this.isConstructor(constructorOrObject)) {
+				if (!Array.isArray(plain)) {
+					constructor = (constructorOrObject as Function)(plain, ...plainParents) as ConstructorType;
+					result = new constructor();
+				}
+			} else {
+				constructor = constructorOrObject as ConstructorType;
+				result = new constructor();
+			}
 		} else {
 			if (constructorOrObject) {
 				result = constructorOrObject;
@@ -40,16 +49,18 @@ export class NaniumObject<T> {
 		}
 
 		// return plain object, if constructor is unknown
-		if (!constructor) {
-			return plain;
+		if (!constructor && !Array.isArray(plain)) {
+			return deepClone ? this.cloneDeep(plain) : plain;
 		}
 
 		// get generic type info
-		globalGenericTypes = globalGenericTypes ?? constructor[genericTypesSymbol];
+		if (constructor) {
+			globalGenericTypes = globalGenericTypes ?? constructor[genericTypesSymbol];
+		}
 
 		// array
 		if (Array.isArray(plain)) {
-			return plain.map(item => this.initObjectCore(item, constructor, globalGenericTypes, undefined, strict));
+			return plain.map(item => this.initObjectCore(item, constructor ?? constructorOrObject, globalGenericTypes, undefined, strict, deepClone, plain));
 		}
 
 		// simple Type or Date
@@ -74,7 +85,7 @@ export class NaniumObject<T> {
 
 		const propertyInfo: NaniumPropertyInfo = constructor[propertyInfoSymbol];
 		let pi: NaniumPropertyInfoCore;
-		let c: ConstructorOrGenericTypeId;
+		let c: ConstructorOrGenericTypeIdOrFkt;
 		for (const property in plain) {
 			if (plain.hasOwnProperty(property)) {
 				if (propertyInfo?.hasOwnProperty(property)) {
@@ -82,41 +93,49 @@ export class NaniumObject<T> {
 					c = pi.ctor ??
 						(localGenericTypes ?? {})[pi.genericTypeId] ??
 						(globalGenericTypes ? globalGenericTypes[pi.genericTypeId] : undefined);
+					if (typeof c === 'function' && !this.isConstructor(c)) {
+						c = (c as Function)(plain[property], plain, ...plainParents);
+					}
 					if (typeof c === 'string') {
 						c = (localGenericTypes ?? {})[pi.genericTypeId] ?? globalGenericTypes[c];
 					}
 					if (c === undefined) {
 						if (!strict) {
-							result[property] = plain[property];
+							result[property] = deepClone ? this.cloneDeep(plain[property]) : plain[property];
 						}
 						continue;
 					}
 					if (c === Object && !pi.localGenerics) { // @Type(Object) - takes the whole object as it is, even in strict mode, because it is explicitly marked as Object/any
-						result[property] = plain[property];
+						result[property] = deepClone ? this.cloneDeep(plain[property]) : plain[property];
 					} else if (c === Array) {
 						if (!Array.isArray(plain[property]) && plain[property] !== undefined && plain[property] !== null) {
-							result[property] = [this.initObjectCore(plain[property], pi.localGenerics as ConstructorType, globalGenericTypes, pi.localGenerics, strict)];
+							result[property] = [this.initObjectCore(plain[property], pi.localGenerics as ConstructorType, globalGenericTypes, pi.localGenerics, strict, deepClone, [plain, ...plainParents])];
 						} else {
-							result[property] = this.initObjectCore(plain[property], pi.localGenerics as ConstructorType, globalGenericTypes, pi.localGenerics, strict);
+							result[property] = this.initObjectCore(plain[property], pi.localGenerics as ConstructorType, globalGenericTypes, pi.localGenerics, strict, deepClone, [plain, ...plainParents]);
 						}
+					} else if (!this.isConstructor(c)) {
+						result[property] = this.initObjectCore(plain[property], (c as Function)(plain, ...plainParents) as ConstructorType, globalGenericTypes, pi.localGenerics, strict, deepClone, [plain, plainParents]);
 					} else {
-						result[property] = this.initObjectCore(plain[property], c as ConstructorType, globalGenericTypes, pi.localGenerics, strict);
+						result[property] = this.initObjectCore(plain[property], c as ConstructorType, globalGenericTypes, pi.localGenerics, strict, deepClone, [plain, plainParents]);
 					}
 				} else {
 					if (typeof localGenericTypes === 'function') { // indexer Properties
+						if (!this.isConstructor(localGenericTypes)) {
+							localGenericTypes = (localGenericTypes as Function)(plain[property], plain, ...plainParents);
+						}
 						if (Object.prototype.hasOwnProperty.call(plain, property)) {
-							result[property] = this.initObjectCore(plain[property], localGenericTypes as ConstructorType, globalGenericTypes, undefined, strict);
+							result[property] = this.initObjectCore(plain[property], localGenericTypes as ConstructorType, globalGenericTypes, undefined, strict, deepClone, [plain, ...plainParents]);
 						}
 						// } else if (typeof plain[property] === 'object' && plain[property] !== null) {
 						// 	if (!Array.isArray(plain[property]) || (plain[property].length && typeof plain[property][0] === 'object')) {
 						// 		if (!strict) {
-						// 			result[property] = plain[property];
+						// 			result[property] = deepClone ? this.cloneDeep(plain[property]) : plain[property];
 						// 			Nanium.logger.warn(`NaniumObject: no type given for property ${property} of class ${constructor.name}`);
 						// 		}
 						// 	}
 					} else {
 						if (!strict) {
-							result[property] = plain[property];
+							result[property] = deepClone ? this.cloneDeep(plain[property]) : plain[property];
 							if (constructor.name !== 'Object') {
 								Nanium.logger.warn(`NaniumObject: no type given for property ${property} of class ${constructor.name}`);
 							}
@@ -131,12 +150,13 @@ export class NaniumObject<T> {
 	static init<T>(dst: T, src: object): void;
 	static init<T>(dst: T, src: object, strict: boolean): void;
 	static init<T>(dst: T, src: object, genericTypes: NaniumGenericTypeInfo): void;
-	static init<T>(dst: T, src: object, genericTypes: NaniumGenericTypeInfo, strict: boolean): void;
+	static init<T>(dst: T, src: object, genericTypes: NaniumGenericTypeInfo, strict: boolean, deepClone?: boolean): void;
 	static init<T>(
 		dst: T,
 		src: object,
 		genericTypesOrStrict?: NaniumGenericTypeInfo | boolean,
-		strict?: boolean
+		strict?: boolean,
+		deepClone?: boolean
 	): void {
 		let genericTypes: NaniumGenericTypeInfo;
 		if (typeof genericTypesOrStrict === 'boolean') {
@@ -145,18 +165,19 @@ export class NaniumObject<T> {
 			strict = strict !== undefined ? strict : this.strictDefault;
 			genericTypes = genericTypesOrStrict;
 		}
-		NaniumObject.initObjectCore<T>(src, dst, genericTypes, undefined, strict);
+		NaniumObject.initObjectCore<T>(src, dst, genericTypes, undefined, strict, deepClone);
 	}
 
-	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeId): T;
-	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeId, strict: boolean): T;
-	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeId, parentCtor: ConstructorType, strict?: boolean): T;
-	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeId, genericTypes: NaniumGenericTypeInfo, strict?: boolean): T;
+	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeIdOrFkt): T;
+	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeIdOrFkt, strict: boolean): T;
+	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeIdOrFkt, parentCtor: ConstructorType, strict?: boolean, deepClone?: boolean): T;
+	static create<T>(src: Partial<T>, ctor: ConstructorOrGenericTypeIdOrFkt, genericTypes: NaniumGenericTypeInfo, strict?: boolean, deepClone?: boolean): T;
 	static create<T>(
 		src: Partial<T>,
-		ctor: ConstructorOrGenericTypeId,
+		ctor: ConstructorOrGenericTypeIdOrFkt,
 		parentConstructorOrGenericTypesInfoOrStrict?: ConstructorType | NaniumGenericTypeInfo | boolean,
 		strict?: boolean,
+		deepClone?: boolean,
 	): T {
 		if (typeof ctor === 'string') {
 			return this.initObjectCore<T>(
@@ -164,8 +185,12 @@ export class NaniumObject<T> {
 				parentConstructorOrGenericTypesInfoOrStrict[genericTypesSymbol] ? parentConstructorOrGenericTypesInfoOrStrict[genericTypesSymbol][ctor] : undefined,
 				parentConstructorOrGenericTypesInfoOrStrict[genericTypesSymbol],
 				undefined,
-				strict);
+				strict,
+				deepClone);
 		} else {
+			if (ctor && !this.isConstructor(ctor as Function)) {
+				ctor = (ctor as Function)(src);
+			}
 			strict = typeof parentConstructorOrGenericTypesInfoOrStrict === 'boolean'
 				? parentConstructorOrGenericTypesInfoOrStrict
 				: strict !== undefined ? strict : this.strictDefault;
@@ -177,10 +202,11 @@ export class NaniumObject<T> {
 			}
 			return this.initObjectCore<T>(
 				src,
-				ctor,
+				ctor as ConstructorType,
 				globalGenericTypes,
 				undefined,
-				strict);
+				strict,
+				deepClone);
 		}
 	}
 
@@ -206,6 +232,36 @@ export class NaniumObject<T> {
 
 		core(obj, fn, []);
 	}
+
+	private static cloneDeep(source: any): any {
+		if (source === undefined) {
+			return undefined;
+		}
+		if (source === null) {
+			return null;
+		}
+		if (Array.isArray(source)) {
+			const target = [];
+			for (const item of source) {
+				target.push(this.cloneDeep(item));
+			}
+			return target;
+		} else if (source instanceof Date) {
+			return new Date(source);
+		} else if (typeof source === 'object') {
+			const target = new source.constructor();
+			for (const key of Object.keys(source)) {
+				target[key] = this.cloneDeep(source[key]);
+			}
+			return target;
+		} else {
+			return source;
+		}
+	}
+
+	private static isConstructor(obj) {
+		return !!obj.prototype && !!obj.prototype.constructor.name;
+	}
 }
 
 export class NaniumPropertyInfo {
@@ -214,9 +270,9 @@ export class NaniumPropertyInfo {
 
 export class NaniumPropertyInfoCore {
 	constructor(
-		public ctor: ConstructorType,
+		public ctor: ConstructorType | ConstructorGetter,
 		public genericTypeId?: string,
-		public localGenerics?: LocalGenerics | ConstructorType
+		public localGenerics?: LocalGenerics | ConstructorType | ConstructorGetter
 	) {
 	}
 }
@@ -240,16 +296,19 @@ export class NaniumEventInfo {
 
 /**
  * Make type information for a property available at runtime.
- * @param clazzOrGenericTypeId If the type of the property is a generic Type, use a unique ID string for this Type. If it is not a generic Type, use the class/constructor. If it is, a dictionary use class Object.
- * @param generics If the property is not a generic Type, but it uses a generic Type (e.g. Stuff<T1, T2>), specify a dictionary with the IDs of the generic types as keys and the constructors as values
+ * @param first If the type of the property is a generic Type, use a unique ID string for this Type. If it is not a generic Type, use the class/constructor. If it is, a dictionary use class Object or Function that returns class objects (parent object is the parameter).
+ * @param second If the property is not a generic Type, but it uses a generic Type (e.g. Stuff<T1, T2>), specify a dictionary with the IDs of the generic types as keys and the constructors as values
  */
-export function Type(clazzOrGenericTypeId: ConstructorOrGenericTypeId, generics?: LocalGenerics | ConstructorType): Function {
+export function Type(
+	first: ConstructorOrGenericTypeIdOrFkt,
+	second?: LocalGenerics | ConstructorType | ConstructorGetter
+): Function {
 	return (target: new () => any, propertyKey: string) => {
 		target.constructor[propertyInfoSymbol] = target.constructor[propertyInfoSymbol] ?? {};
-		if (typeof clazzOrGenericTypeId === 'string') {
-			target.constructor[propertyInfoSymbol][propertyKey] = new NaniumPropertyInfoCore(undefined, clazzOrGenericTypeId, generics);
+		if (typeof first === 'string') {
+			target.constructor[propertyInfoSymbol][propertyKey] = new NaniumPropertyInfoCore(undefined, first, second);
 		} else {
-			target.constructor[propertyInfoSymbol][propertyKey] = new NaniumPropertyInfoCore(clazzOrGenericTypeId, undefined, generics);
+			target.constructor[propertyInfoSymbol][propertyKey] = new NaniumPropertyInfoCore(first, undefined, second);
 		}
 	};
 }
@@ -275,6 +334,10 @@ export type ConstructorType<T = any> = (new (...data: any[]) => T);
 
 export type ConstructorOrGenericTypeId = (ConstructorType | string);
 
+export type ConstructorGetter = ((...parents: Object[]) => ConstructorType);
+
+export type ConstructorOrGenericTypeIdOrFkt = (ConstructorOrGenericTypeId | ConstructorGetter);
+
 export interface LocalGenerics {
-	[genericTypeId: string]: ConstructorOrGenericTypeId;
+	[genericTypeId: string]: ConstructorOrGenericTypeIdOrFkt;
 }